requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-image>=0.20.0
perlin-noise>=1.12
flask>=2.3.0
highlight_text>=0.4.0
